Started Fixing Camera Clipping
Refactored Mario's camera logic completely, so it now adjusts its distance to the mesh if anything obstructs the view. The new implementation, as of now, still has a couple of issues which I intend on fixing in the future - most noticeably, the camera rotation is now much more jagged and jarring.