NPC Walk Animation / Deletion Bug Fixed / Final Tweaks
Added a non-skeletal walking animation to both BobOmbs and Goombas. This required dividing the meshes into sperate chunks (to move separately during animations) and making each feet move along a triangle at an adequate pace (so each foot moves back and forth but takes turns moving vertically). Doing this task also led me to adjust multiple visual values of both NPCs (the most obvious being the between their feet and the ground).

Once with that done, I finally fixed the bug that would sometimes lead to puzzling crashes during NPC deaths. I realized this bug happened due to GameObjects sometimes not communicating with their respective parent properly on deletion, leading to a nullptr exception. Although the logic OverlordEngine has for object deletion should be enough to avoid this type of errors, I ended up adding a second layer of safe-deletion for good measure and it did solve the problem.

Lastly, I did some minor general tweaks to wrap up the project. Some examples of these tweaks were: readjusting the paths of some  of the Cannonballs to avoid clipping through walls; substituting the Cannonball's mesh for a properly unwrapped sphere (both to support a highlight texture and to cast shadows); and calling the Reset function of all objects on Initialize, for good measure, to name a few.

(I also updated both the build, the screenshots and the ReadMe file)